Office League - Input validation (#174)

+++ src/main/resources/services/graphql/graphql-root-mutation.js
@@ -20,9 +20,7 @@ exports.rootMutationType = graphQlLib.createObjectType({
             data: function (env) {
                 var currentPlayerId = getCurrentPlayerId();
                 var adminPlayerIds = env.args.adminPlayerIds || [];
-                if (!currentPlayerId || (adminPlayerIds.indexOf(currentPlayerId) === -1)) {
-                    throw "The current player must be set as admin for a new league.";
-                }
+                checkCreateLeaguePermissions(currentPlayerId, adminPlayerIds);
 
                 var createdLeague = storeLib.createLeague({
                     name: env.args.name,
@@ -48,20 +46,7 @@ exports.rootMutationType = graphQlLib.createObjectType({
                 adminPlayerIds: graphQlLib.list(graphQlLib.GraphQLID)
             },
             data: function (env) {
-                var currentPlayerId = getCurrentPlayerId();
-                var league = storeLib.getLeagueById(env.args.id);
-                if (!league) {
-                    return null;
-                }
-
-                if (!currentPlayerId || (league.adminPlayerIds.indexOf(currentPlayerId) === -1)) {
-                    throw "The league '" + league.name + "' cannot be updated. The current user is not an admin for the league.";
-                }
-
-                var adminPlayerIds = env.args.adminPlayerIds;
-                if (adminPlayerIds && (adminPlayerIds.indexOf(currentPlayerId) === -1)) {
-                    throw "The current player must be set as admin for the league.";
-                }
+                checkUpdateLeaguePermissions(env.args.id, env.args.adminPlayerIds);
 
                 var updatedLeague = storeLib.updateLeague({
                     leagueId: env.args.id,
@@ -80,6 +65,8 @@ exports.rootMutationType = graphQlLib.createObjectType({
                 name: graphQlLib.nonNull(graphQlLib.GraphQLString)
             },
             data: function (env) {
+                checkDeleteLeaguePermissions(env.args.name);
+
                 var deletedId = storeLib.deleteLeagueByName(env.args.name);
                 storeLib.refresh();
                 return deletedId;
@@ -96,12 +83,8 @@ exports.rootMutationType = graphQlLib.createObjectType({
             },
             data: function (env) {
                 var userKey = getCurrentUserKey();
-                if (!userKey) {
-                    throw "Player cannot be created, no logged in user to link to.";
-                }
-                if (isUserLinkedToPlayer(userKey)) {
-                    throw "Player cannot be created, logged in user already linked to another player.";
-                }
+                checkCreatePlayerPermissions(userKey)
+
                 var createdPlayer = storeLib.createPlayer({
                     name: env.args.name,
                     fullname: env.args.fullname,
@@ -125,6 +108,8 @@ exports.rootMutationType = graphQlLib.createObjectType({
                 description: graphQlLib.GraphQLString
             },
             data: function (env) {
+                checkUpdatePlayerPermissions(env.args.id);
+
                 var updatedPlayer = storeLib.updatePlayer({
                     playerId: env.args.id,
                     name: env.args.name,
@@ -145,6 +130,8 @@ exports.rootMutationType = graphQlLib.createObjectType({
                 playerIds: graphQlLib.nonNull(graphQlLib.list(graphQlLib.GraphQLID))
             },
             data: function (env) {
+                checkCreateTeamPermissions(env.args.playerIds);
+
                 var createdTeam = storeLib.createTeam({
                     name: env.args.name,
                     description: env.args.description,
@@ -162,6 +149,8 @@ exports.rootMutationType = graphQlLib.createObjectType({
                 description: graphQlLib.GraphQLString
             },
             data: function (env) {
+                checkUpdateTeamPermissions(env.args.id);
+
                 var updatedTeam = storeLib.updateTeam({
                     teamId: env.args.id,
                     name: env.args.name,
@@ -179,6 +168,8 @@ exports.rootMutationType = graphQlLib.createObjectType({
                 rating: graphQlLib.GraphQLInt
             },
             data: function (env) {
+                checkJoinPlayerLeaguePermissions(env.args.leagueId);
+
                 var createdLeaguePlayer = storeLib.joinPlayerLeague(env.args.leagueId, env.args.playerId, env.args.rating);
                 storeLib.refresh();
                 return createdLeaguePlayer;
@@ -192,6 +183,8 @@ exports.rootMutationType = graphQlLib.createObjectType({
                 ratingDelta: graphQlLib.nonNull(graphQlLib.GraphQLInt)
             },
             data: function (env) {
+                checkUpdatePlayerLeagueRatingPermissions();
+
                 var updatedLeaguePlayer = storeLib.updatePlayerLeagueRating(env.args.leagueId, env.args.playerId, env.args.ratingDelta);
                 storeLib.refresh();
                 return updatedLeaguePlayer;
@@ -205,6 +198,8 @@ exports.rootMutationType = graphQlLib.createObjectType({
                 rating: graphQlLib.GraphQLInt
             },
             data: function (env) {
+                checkJoinTeamLeaguePermissions(env.args.leagueId);
+
                 var createdLeagueTeam = storeLib.joinTeamLeague(env.args.leagueId, env.args.teamId, env.args.rating);
                 storeLib.refresh();
                 return createdLeagueTeam;
@@ -218,6 +213,8 @@ exports.rootMutationType = graphQlLib.createObjectType({
                 ratingDelta: graphQlLib.nonNull(graphQlLib.GraphQLInt)
             },
             data: function (env) {
+                checkUpdateTeamLeagueRatingPermissions();
+
                 var updatedLeagueTeam = storeLib.updateTeamLeagueRating(env.args.leagueId, env.args.teamId, env.args.ratingDelta);
                 storeLib.refresh();
                 return updatedLeagueTeam;
@@ -231,6 +228,8 @@ exports.rootMutationType = graphQlLib.createObjectType({
                 gamePlayers: graphQlLib.nonNull(graphQlLib.list(graphQlInputTypesLib.gamePlayerCreationType))
             },
             data: function (env) {
+                checkCreateGamePermissions(env.args.leagueId, env.args.gamePlayers);
+
                 var createGameParams = storeLib.generateCreateGameParams({
                     leagueId: env.args.leagueId,
                     points: env.args.points || [],
@@ -257,6 +256,8 @@ exports.rootMutationType = graphQlLib.createObjectType({
                 gamePlayers: graphQlLib.nonNull(graphQlLib.list(graphQlInputTypesLib.gamePlayerCreationType))
             },
             data: function (env) {
+                checkUpdateGamePermissions(env.args.gameId, env.args.gamePlayers);
+
                 var createGameParams = storeLib.generateCreateGameParams({
                     leagueId: '',
                     points: env.args.points || [],
@@ -287,6 +288,8 @@ exports.rootMutationType = graphQlLib.createObjectType({
                 id: graphQlLib.nonNull(graphQlLib.GraphQLID)
             },
             data: function (env) {
+                checkDeleteGamePermissions(env.args.id);
+
                 var deletedId = storeLib.deleteGameById(env.args.id);
                 storeLib.refresh();
                 return deletedId;
@@ -316,6 +319,230 @@ exports.rootMutationType = graphQlLib.createObjectType({
     }
 });
 
+var checkCreateLeaguePermissions = function (currentPlayerId, adminPlayerIds) {
+    if (isAdmin()) {
+        return;
+    }
+    if (!currentPlayerId || (adminPlayerIds.indexOf(currentPlayerId) === -1)) {
+        throw "The current player must be set as admin for a new league.";
+    }
+};
+
+var checkUpdateLeaguePermissions = function (leagueId, adminPlayerIds) {
+    if (isAdmin()) {
+        return;
+    }
+
+    var currentPlayerId = getCurrentPlayerId();
+    var league = storeLib.getLeagueById(leagueId);
+    if (!league) {
+        return;
+    }
+
+    if (!currentPlayerId || (league.adminPlayerIds.indexOf(currentPlayerId) === -1)) {
+        throw "The league '" + league.name + "' cannot be updated. The current user is not an admin for the league.";
+    }
+
+    if (adminPlayerIds && (adminPlayerIds.indexOf(currentPlayerId) === -1)) {
+        throw "The current player must be set as admin for the league.";
+    }
+};
+
+var checkDeleteLeaguePermissions = function (leagueName) {
+    if (isAdmin()) {
+        return;
+    }
+
+    var currentPlayerId = getCurrentPlayerId();
+    var league = storeLib.getLeagueByName(env.args.name);
+    if (!league) {
+        return null;
+    }
+    if (!currentPlayerId || (league.adminPlayerIds.indexOf(currentPlayerId) === -1)) {
+        throw "The league '" + league.name + "' cannot be deleted. The current user is not an admin for the league.";
+    }
+};
+
+var checkUpdatePlayerPermissions = function (playerId) {
+    if (isAdmin()) {
+        return;
+    }
+
+    var userKey = getCurrentUserKey();
+    if (!userKey) {
+        throw "Player cannot be updated, no logged in user.";
+    }
+    var player = storeLib.getPlayerById(playerId);
+    if (!player) {
+        return null;
+    }
+    if (player.userKey !== userKey) {
+        throw "User not authorized to update player.";
+    }
+};
+
+var checkCreatePlayerPermissions = function (userKey) {
+    if (!userKey) {
+        throw "Player cannot be created, no logged in user to link to.";
+    }
+    if (isUserLinkedToPlayer(userKey)) {
+        throw "Player cannot be created, logged in user already linked to another player.";
+    }
+};
+
+var checkCreateTeamPermissions = function (playerIds) {
+    if (isAdmin()) {
+        return;
+    }
+
+    var userKey = getCurrentUserKey();
+    if (!userKey) {
+        throw "Team cannot be created, no logged in user.";
+    }
+    var players = storeLib.getPlayersById(playerIds);
+    if (!players || players.length !== 2) {
+        return null;
+    }
+
+    if ((players[0].userKey !== userKey) && (players[1].userKey !== userKey)) {
+        throw "User not authorized to create team.";
+    }
+};
+
+var checkUpdateTeamPermissions = function (teamId) {
+    if (isAdmin()) {
+        return;
+    }
+
+    var userKey = getCurrentUserKey();
+    if (!userKey) {
+        throw "Team cannot be updated, no logged in user.";
+    }
+    var team = storeLib.getTeamById(teamId);
+    if (!team) {
+        return null;
+    }
+    var players = storeLib.getPlayersById(team.playerIds);
+    if (!players || players.length !== 2) {
+        return null;
+    }
+    if ((players[0].userKey !== userKey) && (players[1].userKey !== userKey)) {
+        throw "User not authorized to update team.";
+    }
+};
+
+var checkJoinPlayerLeaguePermissions = function (leagueId) {
+    if (isAdmin()) {
+        return;
+    }
+
+    var userKey = getCurrentUserKey();
+    if (!userKey) {
+        throw "League cannot be updated, no logged in user.";
+    }
+    var league = storeLib.getLeagueById(leagueId);
+    if (!league) {
+        return null;
+    }
+    var adminPlayers = storeLib.getPlayersById(league.adminPlayerIds) || [];
+    var userIsLeagueAdmin = false;
+    adminPlayers.forEach(function (p) {
+        if (p.userKey === userKey) {
+            userIsLeagueAdmin = true;
+        }
+    });
+    if (userIsLeagueAdmin) {
+        throw "User not authorized to update league.";
+    }
+};
+
+var checkUpdatePlayerLeagueRatingPermissions = function () {
+    if (isNotAdmin()) {
+        throw "User not authorized to update league ranking.";
+    }
+};
+
+var checkJoinTeamLeaguePermissions = function (leagueId) {
+    if (isAdmin()) {
+        return;
+    }
+
+    var userKey = getCurrentUserKey();
+    if (!userKey) {
+        throw "League cannot be updated, no logged in user.";
+    }
+    var league = storeLib.getLeagueById(leagueId);
+    if (!league) {
+        return;
+    }
+    var adminPlayers = storeLib.getPlayersById(league.adminPlayerIds) || [];
+    var userIsLeagueAdmin = false;
+    adminPlayers.forEach(function (p) {
+        if (p.userKey === userKey) {
+            userIsLeagueAdmin = true;
+        }
+    });
+    if (userIsLeagueAdmin) {
+        throw "User not authorized to update league.";
+    }
+};
+
+var checkUpdateTeamLeagueRatingPermissions = function () {
+    if (isNotAdmin()) {
+        throw "User not authorized to update league ranking.";
+    }
+};
+
+var checkCreateGamePermissions = function (leagueId, gamePlayers) {
+    var playerIds = gamePlayers.map(function (gp) {
+        return gp.playerId;
+    });
+    var leaguePlayerResp = storeLib.getLeaguePlayersByLeagueIdAndPlayerIds(leagueId, playerIds);
+    if (leaguePlayerResp.count !== playerIds.length) {
+        throw "Game cannot be created, some of the players are not members of the league.";
+    }
+};
+
+var checkUpdateGamePermissions = function (gameId, gamePlayers) {
+    var game = storeLib.getGameById(gameId);
+    if (!game) {
+        return;
+    }
+    var leagueId = game.leagueId;
+    var playerIds = gamePlayers.map(function (gp) {
+        return gp.playerId;
+    });
+    var leaguePlayerResp = storeLib.getLeaguePlayersByLeagueIdAndPlayerIds(leagueId, playerIds);
+    if (leaguePlayerResp.count !== playerIds.length) {
+        throw "Game cannot be updated, some of the players are not members of the league.";
+    }
+};
+
+var checkDeleteGamePermissions = function (gameId) {
+    if (isAdmin()) {
+        return;
+    }
+    var game = storeLib.getGameById(gameId);
+    if (!game) {
+        return;
+    }
+    var playerIds = game.gamePlayers.map(function (gp) {
+        return gp.playerId;
+    });
+    var players = storeLib.getPlayersById(playerIds);
+
+    var userIsGamePlayer = false;
+    var userKey = getCurrentUserKey();
+    players.forEach(function (p) {
+        if (p.userKey === userKey) {
+            userIsGamePlayer = true;
+        }
+    });
+    if (!userIsGamePlayer) {
+        throw "Game cannot be deleted, current user is not a player in the game.";
+    }
+};
+
 var getCurrentPlayerId = function () {
     var user = authLib.getUser();
     if (!user) {
@@ -333,4 +560,12 @@ var getCurrentUserKey = function () {
 var isUserLinkedToPlayer = function (userKey) {
     var player = storeLib.getPlayerByUserKey(userKey);
     return !!player;
+};
+
+var isNotAdmin = function () {
+    return !authLib.hasRole('system.admin');
+};
+
+var isAdmin = function () {
+    return authLib.hasRole('system.admin');
 };
